[add] RECIPE CB RECOM WOOHOO

+++ src/server/src/main/java/dao/RecipeDAO.java
@@ -15,7 +15,7 @@ public interface RecipeDAO {
 	public List<Recipe> findRecipeWithLessThanLevels(float energy, float fat, float proteins, float carbohydrates, String category) throws DAOException;
 	
 
-	public List<Recipe> findRecipeWithLessThanLevelsOrderByPredictionsWithLimit(float energy, float fat, float proteins, float carbohydrates, int limit, int user_id, String category);
+	public List<Recipe> findRecipeWithLessThanLevelsOrderByPredictionsWithLimit(float energy, float fat, float proteins, float carbohydrates, int limit, long user_id, String category);
 
 	public List<Long> getRecipeIdsByCategory(String categoryName) throws DAOException;
 }
